pyproject.toml
[tool.poetry]
name = "masstld"
version = "0.9.2"
description = ""
authors = ["ItsNiceCraft <74358857+ItsNiceCraft@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
dnspython = "^2.5.0"
rich = "^13.7.0"
requests = "^2.31.0"
mdutils = "^1.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
